Business reputation tracker - full pipeline/backend/requirements.txt
flask==3.0.0
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.12.0
watchdog==3.0.0 